requirements.txt
faiss-cpu==1.8.0.post1
numpy==1.26.4
google-generativeai==0.7.0
requests==2.32.3
Pillow==10.3.0
open_clip_torch==2.24.0
torch==2.3.1
pymupdf==1.24.7
ollama==0.2.1
torchvision==0.18.1
matplotlib==3.9.1
transformers==4.42.4
markdownify=0.12.1